refactor(FR-1884): migrate agent table component from webui to bui (#4994)
resolves #4979 ([FR-1884](https://lablup.atlassian.net/browse/FR-1884))

This PR moves UI components from the React package to the backend.ai-ui package to improve reusability:

1. Moved `DoubleTag` to `BAIDoubleTag` in the UI package
2. Moved `BAIIntervalView` to the UI package
3. Moved `BAIProgressWithLabel` to the UI package
4. Added a new `BAIAgentTable` fragment component for displaying agent information

The PR also fixes a z-index issue in the BAITable component where fixed table cells had a z-index that conflicted with the WebUI header, and disables the comma-spacing ESLint rule in the UI package.

**Checklist:**

- [ ] Documentation
- [ ] Minium required manager version
- [ ] Specific setting for review (eg., KB link, endpoint or how to setup)
- [ ] Minimum requirements to check during review
- [ ] Test case(s) to demonstrate the difference of before/after

[FR-1884]: https://lablup.atlassian.net/browse/FR-1884?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
