+++ packages/backend.ai-ui/package.json
@@ -83,6 +83,7 @@
       }
     ],
     "rules": {
+      "comma-spacing": "off",
       "react/react-in-jsx-scope": "off",
       "react/prop-types": "off",
       "import/no-duplicates": "error",
