+++ packages/backend.ai-ui/src/components/index.ts
@@ -82,6 +82,11 @@ export type { UserNodeInList } from './BAIUserNodes';
 export type { BAIUncontrolledInputProps } from './BAIUncontrolledInput';
 export { default as BAIUncontrolledInput } from './BAIUncontrolledInput';
 export { default as BAITag } from './BAITag';
+export { default as BAIIntervalView } from './BAIIntervalView';
+export { default as BAIDoubleTag } from './BAIDoubleTag';
+export type { DoubleTagObjectValue } from './BAIDoubleTag';
+export { default as BAIProgressWithLabel } from './BAIProgressWithLabel';
+export type { BAIProgressWithLabelProps } from './BAIProgressWithLabel';
 export * from './Table';
 export * from './fragments';
 export * from './provider';
