+++ react/src/components/LoginSessionExtendButton.tsx
@@ -1,9 +1,8 @@
 import { useBaiSignedRequestWithPromise } from '../helper';
 import { useSuspenseTanQuery } from '../hooks/reactQueryAlias';
-import BAIIntervalView from './BAIIntervalView';
 import { ClockCircleOutlined } from '@ant-design/icons';
 import { Button, ConfigProvider, Grid, Tooltip } from 'antd';
-import { useUpdatableState, BAIFlex } from 'backend.ai-ui';
+import { useUpdatableState, BAIFlex, BAIIntervalView } from 'backend.ai-ui';
 import { default as dayjs } from 'dayjs';
 import { atom, useAtom } from 'jotai';
 import { Repeat2Icon } from 'lucide-react';
